+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Billing: add profit and profit percentage fields, also add both fields to customer report
 - Billing: add customer report export
 - Comments: Only the creator or admin may edit and delete comments
+- Core: Added "Is empty" filter to custom date filter
 
 24-10-2022: 6.6.123
 - Core: Remember user selected language